Generate a .Rmd file containing code to perform differential expression analysis with the DESeq GLM approach

A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the GLM functionality from the DESeq package. The code is written to a .Rmd file. This function is generally not called by the user, the main interface for performing differential expression analysis is the runDiffExp function.

DESeq.GLM.createRmd(
  data.path,
  result.path,
  codefile,
  sharing.mode,
  disp.method,
  fit.type
)

Arguments

data.path
The path to a .rds file containing the compData object that will be used for the differential expression analysis.

result.path
The path to the file where the result object will be saved.

codefile
The path to the file where the code will be written.

sharing.mode
The method used to select between the individually estimated dispersion and the dispersion estimate obtained by fitting a dispersion-mean relationship to the estimated values for all genes. Possible values are "fit-only" (use the fitted value), "maximum" (take the maximum of the fitted and the estimated value) and "gene-est-only" (use the estimated value).

disp.method
The method used to estimate the dispersion. Possible values are "pooled", "per-condition" and "blind".

fit.type
The fitting method used to get the dispersion-mean relationship. Possible values are "parametric" and "local".

Value

The function generates a .Rmd file containing the code for performing the differential expression analysis. This file can be executed using e.g. the knitr package.

Details

For more information about the methods and the interpretation of the parameters, see the DESeq package and the corresponding publications.

References

Anders S and Huber W (2010): Differential expression analysis for sequence count data. Genome Biology 11:R106

Author

Charlotte Soneson

Examples

try(
if (require(DESeq)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100, 
                                    output.file = file.path(tmpdir, "mydata.rds"))
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "DESeq.GLM", 
           Rmdfunction = "DESeq.GLM.createRmd", 
           output.directory = tmpdir, sharing.mode = "maximum",
           disp.method = "pooled", fit.type = "parametric")
})

#> Loading required package: DESeq

#> Warning: there is no package called ‘DESeq’