Generate a .Rmd file containing code to perform differential expression analysis with NBPSeq

A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using NBPSeq. The code is written to a .Rmd file. This function is generally not called by the user, the main interface for performing differential expression analysis is the runDiffExp function.

NBPSeq.createRmd(data.path, result.path, codefile, norm.method, disp.method)

Arguments

data.path
The path to a .rds file containing the compData object that will be used for the differential expression analysis.

result.path
The path to the file where the result object will be saved.

codefile
The path to the file where the code will be written.

norm.method
The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. The normalization factors are calculated using the calcNormFactors function from the edgeR package. Possible values are "TMM", "RLE", "upperquartile" and "none".

disp.method
The method to use to estimate the dispersion values. Possible values are "NBP" and "NB2".

Value

The function generates a .Rmd file containing the code for performing the differential expression analysis. This file can be executed using e.g. the knitr package.

Details

For more information about the methods and the interpretation of the parameters, see the NBPSeq and edgeR packages and the corresponding publications.

References

Robinson MD, McCarthy DJ and Smyth GK (2010): edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139-140

Robinson MD and Oshlack A (2010): A scaling normalization method for differential expression analysis of RNA-seq data. Genome Biology 11:R25

Di Y, Schafer DW, Cumbie JS, and Chang JH (2011): The NBP Negative Binomial Model for Assessing Differential Gene Expression from RNA-Seq. Statistical Applications in Genetics and Molecular Biology 10(1), 1-28

Author

Charlotte Soneson

Examples

try(
if (require(NBPSeq)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100, 
                                    output.file = file.path(tmpdir, "mydata.rds"))
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "NBPSeq", 
           Rmdfunction = "NBPSeq.createRmd", 
           output.directory = tmpdir, norm.method = "TMM", disp.method = "NBP")
})

#> Loading required package: NBPSeq

#> 
#> 
#> processing file: tempcode.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: unnamed-chunk-1 (with options) 
#> List of 6
#>  $ echo   : logi TRUE
#>  $ eval   : logi TRUE
#>  $ include: logi TRUE
#>  $ message: logi FALSE
#>  $ error  : logi TRUE
#>  $ warning: logi TRUE
#> 

#> Loading required package: edgeR

#> Loading required package: limma

#> 
#> Attaching package: 'limma'

#> The following object is masked from 'package:DESeq2':
#> 
#>     plotMA

#> The following object is masked from 'package:BiocGenerics':
#> 
#>     plotMA

#> 

#> output file: tempcode.md

#> [1] TRUE