Generate a .Rmd file containing code to perform differential expression analysis with baySeq

A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the baySeq package. The code is written to a .Rmd file. This function is generally not called by the user, the main interface for performing differential expression analysis is the runDiffExp function.

baySeq.createRmd(
  data.path,
  result.path,
  codefile,
  norm.method,
  equaldisp,
  sample.size = 5000,
  estimation = "QL",
  pET = "BIC"
)

Arguments

data.path
The path to a .rds file containing the compData object that will be used for the differential expression analysis.

result.path
The path to the file where the result object will be saved.

codefile
The path to the file where the code will be written.

norm.method
The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. Possible values are "quantile", "total" and "edgeR".

equaldisp
Logical parameter indicating whether or not equal dispersion should be assumed across all conditions.

sample.size
The size of the sample used to estimate the priors (default 5000).

estimation
The approach used to estimate the priors. Possible values are "QL" (default), "ML" and "edgeR".

pET
The method used to re-estimate the priors. Possible values are "BIC" (default), "none" and "iteratively".

Value

The function generates a .Rmd file containing the code for performing the differential expression analysis. This file can be executed using e.g. the knitr package.

Details

For more information about the methods and the interpretation of the parameters, see the baySeq package and the corresponding publications.

References

Hardcastle TJ (2012): baySeq: Empirical Bayesian analysis of patterns of differential expression in count data. R package

Hardcastle TJ and Kelly KA (2010): baySeq: Empirical Bayesian methods for identifying differential expression in sequence count data. BMC Bioinformatics 11:422

Examples

try(
if (require(baySeq)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000,
                                    samples.per.cond = 5, n.diffexp = 100,
                                    output.file = file.path(tmpdir, "mydata.rds"))
## Note! In the interest of speed, we set sample.size=10 in this example. 
## In a real analysis, much larger sample sizes are recommended (the default is 5000).
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "baySeq",
           Rmdfunction = "baySeq.createRmd",
           output.directory = tmpdir, norm.method = "edgeR",
           equaldisp = TRUE, sample.size = 10)
})

#> 
#> 
#> processing file: tempcode.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: unnamed-chunk-1 (with options) 
#> List of 6
#>  $ echo   : logi TRUE
#>  $ eval   : logi TRUE
#>  $ include: logi TRUE
#>  $ message: logi TRUE
#>  $ error  : logi TRUE
#>  $ warning: logi TRUE
#> 
#> 

#> output file: tempcode.md

#> [1] TRUE