Check the validity of a phyloCompData object

Check the validity of a phyloCompData object. An object that passes the check can be used as the input for the differential expression analysis methods interfaced by compcodeR.

check_phyloCompData(object)

Arguments

object

A phyloCompData object

Author

Charlotte Soneson, Paul Bastide

Examples

mydata <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                samples.per.cond = 5, n.diffexp = 100,
                                id.species = factor(1:10),
                                tree = ape::rphylo(10, 1, 0),
                                lengths.relmeans = "auto", lengths.dispersions = "auto")
#> Warning: `id.species` is not named. I'm naming them, assuming they are in the same order as the tree.
check_phyloCompData(mydata)
#> [1] TRUE