Class compData

The compData class is used to store information about the experiment, such as the count matrix, sample and variable annotations, information regarding the generation of the data and results from applying a differential expression analysis to the data.

Slots

count.matrix:

The read count matrix, with genes as rows and samples as columns. Class matrix

sample.annotations:

A data frame containing sample annotation information for all samples in the data set. Must contain at least a column named condition, encoding the division of the samples into two classes. The row names should be the same as the column names of count.matrix. Class data.frame

info.parameters:

A list of parameters detailing the simulation process used to generate the data. Must contain at least two entries, named dataset (an informative name for the data set/simulation setting) and uID (a unique ID for the specific data set instance). Class list

filtering:

A character string detailing the filtering process that has been applied to the data. Class character

variable.annotations:

Contains information regarding the variables, such as the differential expression status, the true mean, dispersion and effect sizes. If present, the row names should be the same as those of count.matrix. Class data.frame

analysis.date:

(If a differential expression analysis has been performed and the results are included in the compData object). Gives the date when the differential expression analysis was performed. Class character

package.version:

(If a differential expression analysis has been performed and the results are included in the compData object). Gives the version numbers of the package(s) used for the differential expression analysis. Class character

method.names:

(If a differential expression analysis has been performed and the results are included in the compData object). A list, containing the name of the method used for the differential expression analysis. The list should have two entries: full.name and short.name, where the full.name is the full (potentially long) name identifying the method, and short.name may be an abbreviation. Class list

code:

(If a differential expression analysis has been performed and the results are included in the compData object). A character string containing the code that was used to run the differential expression analysis. The code should be in R markdown format. Class character

result.table:

(If a differential expression analysis has been performed and the results are included in the compData object). Contains the results of the differential expression analysis, in the form of a data frame with one row per gene. Must contain at least one column named score, where a higher value corresponds to 'more strongly differentially expressed genes'. Class data.frame

Methods

count.matrix

signature(x="compData")

count.matrix<-

signature(x="compData",value="matrix"): Get or set the count matrix in a compData object. value should be a numeric matrix.

sample.annotations

signature(x="compData")

sample.annotations<-

signature(x="compData",value="data.frame"): Get or set the sample annotations data frame in a compData object. value should be a data frame with at least a column named 'condition'.

info.parameters

signature(x="compData")

info.parameters<-

signature(x="compData",value="list"): Get or set the list with info parameters in a compData object. value should be a list with at least elements named 'dataset' and 'uID'.

filtering

signature(x="compData")

filtering<-

signature(x="compData",value="character"): Get or set the information about the filtering in a compData object. value should be a character string describing the filtering that has been performed.

variable.annotations

signature(x="compData")

variable.annotations<-

signature(x="compData",value="data.frame"): Get or set the variable annotations data frame in a compData object. value should be a data frame.

analysis.date

signature(x="compData")

analysis.date<-

signature(x="compData",value="character"): Get or set the analysis date in a compData object. value should be a character string describing when the differential expression analysis of the data was performed.

package.version

signature(x="compData")

package.version<-

signature(x="compData",value="character"): Get or set the information about the package version in a compData object. value should be a character string detailing which packages and versions were used to perform the differential expression analysis of the data.

method.names

signature(x="compData")

method.names<-

signature(x="compData",value="list"): Get or set the method names in a compData object. value should be a list with slots full.name and short.name, giving the full name and an abbreviation for the method that was used to perform the analysis of the data.

code

signature(x="compData")

code<-

signature(x="compData",value="character"): Get or set the code slot in a compData object. value should be a character string in R markdown format, giving the code that was run to obtain the results from the differential expression analysis.

result.table

signature(x="compData")

result.table<-

signature(x="compData",value="data.frame"): Get or set the result table in a compData object. value should be a data frame with one row per gene, and at least a column named 'score'.

Construction

An object of the class compData can be constructed using the compData function.

Author

Charlotte Soneson