Scale the variances

Scale the variances of the process simulation so that they are equal to log_variance_phylo.

scale_variance_process(
  log_variance_phylo,
  tree,
  model.process,
  selection.strength
)

Arguments

tree

a dated phylogenetic tree of class phylo with `samples.per.cond * 2` species.

model.process

the process to be used for phylogenetic simulations. One of "BM" or "OU", default to "BM".

selection.strength

if the process is "OU", the selection strength parameter.

Value

A matrix N * P of factors to multiply the simulated phylogenetic residuals.