Generate a .Rmd file containing code to normalize data.

Generate a .Rmd file containing code to normalize data.

writeNormalization(
  norm.method,
  length.normalization,
  data.transformation,
  codefile
)

Arguments

norm.method

The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. The normalization factors are calculated using the calcNormFactors of the edgeR package. Possible values are "TMM", "RLE", "upperquartile" and "none"

length.normalization

one of "none" (no correction), "TPM", "RPKM" (default). See details.

data.transformation

one of "log2", "asin(sqrt)" or "sqrt." Data transformation to apply to the normalized data.

codefile

Details

The length.matrix field of the phyloCompData object is used to normalize the counts.

none:

No length normalization.

TPM:

The raw counts are divided by the length of their associated genes before normalization by voom.

RPKM:

The log2 length is substracted to the log2 CPM computed by voom for each gene and sample.