Check the validity of a compData result object

Check the validity of a compData object containing differential expression results. An object that passes the check can be used as the input for the method comparison functions in compcodeR.

check_compData_results(object)

Arguments

object
A compData object

Author

Charlotte Soneson

Examples

tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                samples.per.cond = 5, n.diffexp = 100,
                                output.file = file.path(tmpdir, "mydata.rds"))
## Check an object without differential expression results
check_compData_results(mydata)

#> [1] "Object must contain a list named 'method.names' identifying the differential expression method used."

runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), 
           result.extent = "voom.limma", 
           Rmdfunction = "voom.limma.createRmd", 
           output.directory = tmpdir, norm.method = "TMM")

#> 
#> 
#> processing file: tempcode.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: unnamed-chunk-1 (with options) 
#> List of 6
#>  $ echo   : logi TRUE
#>  $ eval   : logi TRUE
#>  $ include: logi TRUE
#>  $ message: logi FALSE
#>  $ error  : logi TRUE
#>  $ warning: logi TRUE
#> 
#> 

#> output file: tempcode.md

#> [1] TRUE

resdata <- readRDS(file.path(tmpdir, "mydata_voom.limma.rds"))
## Check an object containing differential expression results
check_compData_results(resdata)

#> [1] TRUE